+++ regressions/ck_pr/validate/ck_pr_store.c
@@ -41,7 +41,7 @@
 #define CK_PR_STORE_B(w)								\
 	{										\
 		uint##w##_t t = (uint##w##_t)-1, a = 0, b;				\
-		ck_pr_store_##w(&b, 1ULL << (w - 1));					\
+		ck_pr_store_##w(&b, (uint64_t)1 << (w - 1));				\
 		unsigned int i;								\
 		printf("ck_pr_store_" #w ": ");						\
 		if (w < 10)								\
